hardware-server/requirements.txt
pyserial
websockets